+++ modal/sandbox.py
@@ -538,6 +538,16 @@ class _Sandbox(_Object, type_prefix="sb"):
 
         return self._tunnels
 
+    async def reload_volumes(self) -> None:
+        """Reload all Volumes mounted in the Sandbox."""
+        task_id = await self._get_task_id()
+        await retry_transient_errors(
+            self._client.stub.ContainerReloadVolumes,
+            api_pb2.ContainerReloadVolumesRequest(
+                task_id=task_id,
+            ),
+        )
+
     async def terminate(self) -> None:
         """Terminate Sandbox execution.